# examples/backend/common/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Learning Admin API
  version: 1.0.0
  description: Shared contract implemented by NestJS, FastAPI, and Gin.
servers:
  - url: http://localhost:3001/api
    description: NestJS
  - url: http://localhost:3002/api
    description: FastAPI
  - url: http://localhost:3003/api
    description: Gin
security:
  - bearerAuth: []
paths:
...
  /files/presign:
    post:
      operationId: presignFile
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/FileIntent' } }
      responses:
        '201':
          description: Short-lived direct upload URL.
          content:
            application/json: { schema: { $ref: '#/components/schemas/PresignedUpload' } }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /files/{fileId}/download:
    get:
      operationId: downloadFile
      parameters:
        - { name: fileId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '302':
          description: Redirect to a short-lived signed object URL after authorization.
          headers:
            Location: { schema: { type: string, format: uri } }
        '404': { $ref: '#/components/responses/Problem' }
...
  /orders:
    post:
      operationId: createOrder
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/OrderInput' } }
      responses:
        '201':
          description: Inventory reserved and order created in one transaction.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Order' } }
        '409': { $ref: '#/components/responses/Problem' }
  /payments/callback:
    post:
      operationId: paymentCallback
      security: []
      parameters:
        - { name: X-Signature, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/PaymentCallback' } }
      responses:
        '204': { description: Idempotent callback accepted. }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 128 } }
...
  responses:
...
    Problem:
      description: Stable problem response.
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
...
  schemas:
    Uuid: { type: string, format: uuid }
...
    FileIntent:
      type: object
      additionalProperties: false
      required: [filename, contentType, size, sha256]
      properties:
        filename: { type: string, minLength: 1, maxLength: 255 }
        contentType: { type: string, maxLength: 120 }
        size: { type: integer, minimum: 1, maximum: 52428800 }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    PresignedUpload:
      type: object
      required: [fileId, objectKey, uploadUrl, expiresAt]
      properties:
        fileId: { $ref: '#/components/schemas/Uuid' }
        objectKey: { type: string }
        uploadUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
...
    OrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { $ref: '#/components/schemas/Uuid' }
              quantity: { type: integer, minimum: 1, maximum: 999 }
    Order:
      type: object
      required: [id, status, amount, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        status: { type: string, enum: [pending, paid, cancelled, refunded] }
        amount: { type: string, pattern: '^\d+\.\d{2}$' }
        version: { type: integer, minimum: 1 }
    PaymentCallback:
      type: object
      additionalProperties: false
      required: [provider, providerEventId, orderId, status]
      properties:
        provider: { type: string, minLength: 1, maxLength: 80 }
        providerEventId: { type: string, minLength: 1, maxLength: 190 }
        orderId: { $ref: '#/components/schemas/Uuid' }
        status: { type: string, const: paid }
...
    FieldError:
      type: object
      required: [field, messages]
      properties:
        field: { type: string }
        messages: { type: array, items: { type: string } }
    Problem:
      type: object
      additionalProperties: false
      required: [status, code, detail, requestId]
      properties:
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }
        fields: { type: array, items: { $ref: '#/components/schemas/FieldError' } }